requirements.txt
discord.py
python-dotenv
playwright